Phase 1 Stage 1: deep-agent-compatible agent harness in core
Refactor the lean agent harness so it shares one contract + one streaming
implementation with the upcoming deep agent, without core depending on
deepagents. Prep for @gaunt-sloth/agent's GthDeepAgent.

- types.ts: move AgentStreamEvent here (the shared renderer contract); add
  streamWithEvents/streamWithEventsResume to GthAgentInterface; add the
  GthCompiledGraph structural surface and GthAgentFactory.
- GthAbstractAgent: new base holding all graph-agnostic plumbing
  (invoke/stream/streamWithEvents*/processEventStream/cleanup, client-tool
  interrupt() stubbing, getEffectiveConfig). Operates only on GthCompiledGraph,
  so it imports neither langchain nor deepagents graph builders.
- GthLangChainAgent: now extends GthAbstractAgent; keeps only the createAgent
  -specific init(). Re-exports AgentStreamEvent for back-compat.
- GthAgentRunner: inject an optional agentFactory (defaults to GthLangChainAgent)
  so @gaunt-sloth/agent can plug in GthDeepAgent over the same runner.

Lean path behavior unchanged: build green, 540/540 unit tests pass, lint clean.
Per docs/PLAN.md Phase 1 (revised: lean harness stays in core).
